+++ src/token.rs
@@ -46,27 +46,6 @@ impl std::fmt::Display for Token {
     }
 }
 
-impl Token {
-    pub fn to_value(&self) -> Value {
-        if self.t == TokenType::True {
-            return Value::Bool(true);
-        }
-        else if self.t == TokenType::False {
-            return Value::Bool(false);
-        }
-        else if self.t == TokenType::Null {
-            return Value::Null;
-        }
-        else if self.t == TokenType::Number {
-            return Value::Number(
-                self.val.parse::<f64>()
-                    .expect("Error parsing float! This might be a problem with the interpreter itself.")
-            );
-        }
-        unimplemented!()
-    }
-}
-
 #[derive(Clone, Debug)]
 pub struct TokenPos {
     pub from_col: u32,
